Move SPM manifest to the root so SPM can use it now
